requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
flask_sqlalchemy>=3.0.0
functions-framework==3.*
gunicorn==20.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.3.2
MarkupSafe==2.1.5
matplotlib==3.8.0
numpy==1.26.4
pandas==2.1.4
python-dotenv
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.1
threadpoolctl==3.5.0
Werkzeug==3.0.3
blosc2==2.0.0
FuzzyTM>=0.4.0
ipykernel==6.25.0
ipython==8.15.0
shap
xgboost
lightgbm
google-auth
google-cloud-storage
psycopg2-binary==2.9.3
SQLAlchemy==1.4.36